natural-health-rag/requirements.txt
# Natural Health RAG - Dependencies

# LangChain core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Vector stores
langchain-chroma>=0.1.4        # local, free
# langchain-pinecone>=0.2.0    # uncomment for cloud

# Embeddings (choose one)
# For local free embeddings: install ollama, then run `ollama pull nomic-embed-text`
openai>=1.40.0                 # for OpenAI embeddings + LLM

# Document loading
pypdf>=4.0.0                   # PDF loading
beautifulsoup4>=4.12.0         # HTML scraping
requests>=2.31.0

# UI
streamlit>=1.35.0

# Optional: Pinecone
# pinecone-client>=3.0.0

# Dev
python-dotenv>=1.0.0
